requirements.txt
dj-database-url==1.0.0
dj-static==0.0.6
django==4.2.0
django-environ==0.9.0
django-model-utils==4.2.0
django-widget-tweaks==1.5.0
pytest-django==4.8.0
gunicorn==20.1.0
psycopg2==2.9.9
python-decouple==3.6
python-dotenv==0.21.0
pillow==10.3.0
sqlparse==0.4.3
static3==0.7.0
whitenoise==6.2.0